+++ src/Taxes/Contracts/TaxCalculator.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 
 namespace Vanilo\Taxes\Contracts;
 
+use Vanilo\Adjustments\Contracts\Adjustable;
+use Vanilo\Adjustments\Contracts\Adjuster;
 use Vanilo\Contracts\DetailedAmount;
 use Vanilo\Contracts\Schematized;
 
@@ -21,17 +23,7 @@ interface TaxCalculator extends Schematized
 {
     public static function getName(): string;
 
-    /**
-     * We don't set the return type on the language level
-     * since the adjustments module is optional
-     * @todo Remove this nonsense in v6 and add the proper return type
-     *
-     * @return null|\Vanilo\Adjustments\Contracts\Adjuster
-     */
-    public function getAdjuster(?array $configuration = null): ?object;
+    public function getAdjuster(?array $configuration = null): ?Adjuster;
 
-    /**
-     * @todo change ?object to Adjustable in v6. Also - when can the configuration be null?
-     */
-    public function calculate(?object $subject = null, ?array $configuration = null): DetailedAmount;
+    public function calculate(Adjustable $subject, ?array $configuration = null): DetailedAmount;
 }